+++ CHANGES.rst
@@ -13,12 +13,12 @@ upstream of anybox.recipe.odoo, and will change in the future.
 
 anybox.recipe.odoo 1.9.2 (unreleased)
 -------------------------------------
+- github #18: Support develop-dir option from gp.vcsdevelop
 - Odoo renaming in the doc
 - adapted nightly archive URL
 
 anybox.recipe.odoo 1.9.1 (2014-12-17)
 -------------------------------------
-- github #18: Support develop-dir option from gp.vcsdevelop
 - github #7: template for ``upgrade.py`` was missing for a.r.odoo
 
 anybox.recipe.openerp 1.9.1 (unreleased)